Refactor game.js into modular components with separate concerns
