+++ test_smg/vite.config.js
@@ -5,7 +5,15 @@ import tailwindcss from '@tailwindcss/vite';
 export default defineConfig({
     plugins: [
         laravel({
-            input: ['resources/css/app.css', 'resources/js/app.js', 'resources/js/game.js'],
+            input: [
+                'resources/css/app.css',
+                'resources/js/app.js',
+                'resources/js/game.js',
+                'resources/js/character.js',
+                'resources/js/dice.js',
+                'resources/js/gameState.js',
+                'resources/js/ui.js',
+            ],
             refresh: true,
         }),
         tailwindcss(),
